Devex 974: Move action into different folder that's a little more specific
